Fix remaining merge issues
